ARRISXB3-3439 : CcspPandM patches from RDKB to yocto branch Reason for change: RDKB patches to yocto Test Procedure: Test every thing. Risks: Low

+++ custom/comcast/source/TR-181/custom_include/cosa_x_cisco_com_hotspot_apis.h
@@ -52,7 +52,7 @@ _COSA_DML_HOTSPOT_ASSODEV
 
     char                        MACAddress[18];
     char                        Hostname[257];
-    ULONG                       RSSILevel;
+    int                         RSSILevel;
     char                        IPv4Address[17];
     char                        DHCPv4Status[64];
     char                        IPv6Address[64];
